LPD-94449 Move testIndividualPublishPreservesCircularArticleReference to JournalExportImportTest
